+++ Example/PhoneLogin/Models/SnabbleConfig.swift
@@ -7,21 +7,24 @@
 //
 
 import Foundation
-import SnabbleNetwork
+import SnabblePhoneAuth
 
 extension Configuration {
     static var appId: String {
         "snabble-sdk-demo-app-oguh3x"
     }
+    
     /// A `String`with the project identifier used for backend communication
     public static var projectId: String {
         "snabble-sdk-demo-beem8n"
     }
+    
     static var testing: Self {
         return .init(
             appId: appId,
             appSecret: "BWXJ2BFC2JRKRNW4QBASQCF2TTANPTVPOXQJM57JDIECZJQHZWOQ====",
-            environment: .testing
+            domain: .testing,
+            projectId: Self.projectId
         )
     }
     
@@ -29,7 +32,8 @@ extension Configuration {
         return .init(
             appId: appId,
             appSecret: "P3SZXAPPVAZA5JWYXVKFSGGBN4ZV7CKCWJPQDMXSUMNPZ5IPB6NQ====",
-            environment: .staging
+            domain: .staging,
+            projectId: Self.projectId
         )
     }
 
@@ -37,11 +41,12 @@ extension Configuration {
         return .init(
             appId: appId,
             appSecret: "2TKKEG5KXWY6DFOGTZKDUIBTNIRVCYKFZBY32FFRUUWIUAFEIBHQ====",
-            environment: .production
+            domain: .production,
+            projectId: Self.projectId
         )
     }
 
-    static func config(for environment: Environment) -> Self {
+    static func config(for environment: Domain) -> Self {
         switch environment {
         case .testing:
             return Self.testing