+++ Sources/Network/Models/Configuration.swift
@@ -10,14 +10,15 @@ import Foundation
 public struct Configuration {
     public let appId: String
     public let appSecret: String
-    public let environment: Environment
+    public let domain: Domain
+    public let projectId: String
 
-    public init(appId: String, appSecret: String, environment: Environment) {
+    public init(appId: String, appSecret: String, domain: Domain, projectId: String) {
         self.appId = appId
         self.appSecret = appSecret
-        self.environment = environment
+        self.domain = domain
+        self.projectId = projectId
     }
 }
 
-extension Configuration: Equatable {
-}
+extension Configuration: Equatable {}