+++ Tests/Network/Endpoints/PhoneEndpointTests.swift
@@ -21,18 +21,18 @@ final class PhoneEndpointTests: XCTestCase {
         // Put teardown code here. This method is called after the invocation of each test method in the class.
     }
 
-    var configuration: Configuration = .init(appId: "1", appSecret: "2", environment: .production)
+    var configuration: Configuration = .init(appId: "1", appSecret: "2", domain: .production, projectId: "1")
     var appUser: AppUser = .init(id: "555", secret: "123-456-789")
 
     func testAuth() throws {
         let endpoint = Endpoints.Phone.auth(configuration: configuration, phoneNumber: phoneNumber)
         XCTAssertEqual(endpoint.configuration, configuration)
-        XCTAssertEqual(endpoint.environment, .production)
+        XCTAssertEqual(endpoint.domain, .production)
         XCTAssertEqual(endpoint.method.value, "POST")
-        XCTAssertEqual(endpoint.path, "/1/phone/auth")
+        XCTAssertEqual(endpoint.path, "/1/verification/sms")
         XCTAssertNil(endpoint.token)
         let urlRequest = try endpoint.urlRequest()
-        XCTAssertEqual(urlRequest.url?.absoluteString, "https://api.snabble.io/1/phone/auth")
+        XCTAssertEqual(urlRequest.url?.absoluteString, "https://api.snabble.io/1/verification/sms")
         XCTAssertEqual(urlRequest.httpMethod, "POST")
 
         let data = try! JSONSerialization.data(withJSONObject: [
@@ -44,31 +44,31 @@ final class PhoneEndpointTests: XCTestCase {
     func testLogin() throws {
         let endpoint = Endpoints.Phone.login(configuration: configuration, phoneNumber: phoneNumber, OTP: otp)
         XCTAssertEqual(endpoint.configuration, configuration)
-        XCTAssertEqual(endpoint.environment, .production)
+        XCTAssertEqual(endpoint.domain, .production)
         XCTAssertEqual(endpoint.method.value, "POST")
-        XCTAssertEqual(endpoint.path, "/1/phone/login")
+        XCTAssertEqual(endpoint.path, "/1/verification/sms/otp")
         XCTAssertNil(endpoint.token)
         let urlRequest = try endpoint.urlRequest()
-        XCTAssertEqual(urlRequest.url?.absoluteString, "https://api.snabble.io/1/phone/login")
+        XCTAssertEqual(urlRequest.url?.absoluteString, "https://api.snabble.io/1/verification/sms/otp")
         XCTAssertEqual(urlRequest.httpMethod, "POST")
 
         let data = try! JSONSerialization.data(withJSONObject: [
             "otp": otp,
             "phoneNumber": phoneNumber
         ])
-        XCTAssertEqual(urlRequest.httpBody, data)
+        XCTAssertEqual(urlRequest.httpBody?.count, data.count)
     }
 
     func testDelete() throws {
         let endpoint = Endpoints.Phone.delete(configuration: configuration, phoneNumber: phoneNumber)
         XCTAssertEqual(endpoint.configuration, configuration)
-        XCTAssertEqual(endpoint.environment, .production)
-        XCTAssertEqual(endpoint.method.value, "DELETE")
-        XCTAssertEqual(endpoint.path, "/1/phone/users")
+        XCTAssertEqual(endpoint.domain, .production)
+        XCTAssertEqual(endpoint.method.value, "POST")
+        XCTAssertEqual(endpoint.path, "/1/verification/sms/delete")
         XCTAssertNil(endpoint.token)
         let urlRequest = try endpoint.urlRequest()
-        XCTAssertEqual(urlRequest.url?.absoluteString, "https://api.snabble.io/1/phone/users")
-        XCTAssertEqual(urlRequest.httpMethod, "DELETE")
+        XCTAssertEqual(urlRequest.url?.absoluteString, "https://api.snabble.io/1/verification/sms/delete")
+        XCTAssertEqual(urlRequest.httpMethod, "POST")
 
         let data = try! JSONSerialization.data(withJSONObject: [
             "phoneNumber": phoneNumber