+++ CHANGELOG.md
@@ -4,5 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-### [1.0.0] - 2024-01-30
+## [1.1.0]
+
+### Added
+* Accessible HTTP Errors
+* Default `CountryCallingCode` information
+* Wrapper class to access the API `PhoneAuth`
+
+### Updated 
+* lachlanbell/SwiftOTP 3.0.2 (was 3.0.1)
+
+## [1.0.0] - 2024-01-30
 * Initital release